add cairo's live-demo
