+++ demos/cairo/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cairo_button"
+name = "cairo_frame"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,5 +7,5 @@ edition = "2021"
 
 [dependencies]
 flemish = { path = "../../" }
+fltk = { version = "^1.4", features = ["use-ninja", "cairoext"] }
 cairo-rs = "0.18"
-cairo-blur = "^0.1"
