+++ demos/cairo/README.md
@@ -1,0 +1,5 @@
+# Cairo demo
+
+Use Cairo for custom drawing.
+
+![img](https://github.com/fltk-rs/demos/blob/master/cairo/assets/scrot.png)
