+++ demos/cairo/src/main.rs
@@ -1,38 +1,29 @@
-#![forbid(unsafe_code)]
-
 mod model;
 
 use {
     cairo::{Context, Format, ImageSurface},
     flemish::{
         app,
-        button::Button,
-        color_themes, draw,
-        enums::Event,
-        enums::{Align, Color, ColorDepth, Font, Shortcut},
         frame::Frame,
-        group::Flex,
-        image::RgbImage,
-        menu::{MenuButton, MenuButtonType, MenuFlag},
+        color_themes,
+        enums::{Color, FrameType},
         prelude::*,
-        OnEvent, OnMenuEvent, Sandbox, Settings,
+        OnEvent, Sandbox, Settings,
     },
     model::Model,
 };
 
 #[derive(Clone, Copy)]
 pub enum Message {
-    Inc,
-    Dec,
-    Quit,
+    Change(usize),
 }
 
 fn main() {
     Model::new().run(Settings {
-        size: (640, 360),
+        size: (260, 260),
         ignore_esc_close: true,
         resizable: false,
-        background: Some(Color::from_u32(0xfdf6e3)),
+        background: Some(Color::White),
         color_map: Some(color_themes::TAN_THEME),
         scheme: Some(app::Scheme::Base),
         ..Default::default()
@@ -47,177 +38,60 @@ impl Sandbox for Model {
     }
 
     fn title(&self) -> String {
-        format!("{} - FlCairo", self.value())
+        format!("{} - FlCairo", self.state[0])
     }
 
     fn view(&mut self) {
-        let mut page = Flex::default()
-            .with_size(600, 200)
-            .center_of_parent()
-            .column();
-
-        let hero = Flex::default(); //HERO
-        crate::cairobutton()
-            .with_label("@#<")
-            .on_event(move |_| Message::Dec);
-        crate::frame(&self.value()).handle(crate::popup);
-        crate::cairobutton()
-            .with_label("@#>")
-            .on_event(move |_| Message::Inc);
-        hero.end();
-
-        page.end();
-        page.set_pad(0);
-        page.set_margin(0);
+        fltk::app::cairo::set_autolink_context(true);
+        let mut frame = cairowidget(5, 5, 100, 100, "Box1");
+        frame.set_color(match self.state[0] {
+            true => Color::Red,
+            false => Color::DarkRed,
+        });
+        frame.on_event(move |_| Message::Change(0));
+        let mut frame = cairowidget(80, 80, 100, 100, "Box2");
+        frame.set_color(match self.state[1] {
+            true => Color::Yellow,
+            false => Color::DarkYellow,
+        });
+        frame.on_event(move |_| Message::Change(1));
+        let mut frame = cairowidget(155, 155, 100, 100, "Box3");
+        frame.set_color(match self.state[2] {
+            true => Color::Green,
+            false => Color::DarkGreen,
+        });
+        frame.clone().on_event(move |_| Message::Change(2));
     }
 
     fn update(&mut self, message: Message) {
         match message {
-            Message::Inc => self.inc(),
-            Message::Dec => self.dec(),
-            Message::Quit => app::quit(),
+            Message::Change(idx) => self.change(idx),
         }
     }
 }
 
-fn menu() -> MenuButton {
-    MenuButton::default()
-        .with_type(MenuButtonType::Popup3)
-        .with_label("@#menu")
-        .on_item_event(
-            "@#+  &Increment",
-            Shortcut::Ctrl | 'i',
-            MenuFlag::Normal,
-            move |_| Message::Inc,
-        )
-        .on_item_event(
-            "@#-  &Decrement",
-            Shortcut::Ctrl | 'd',
-            MenuFlag::Normal,
-            move |_| Message::Dec,
-        )
-        .on_item_event(
-            "@#1+  Quit",
-            Shortcut::Ctrl | 'q',
-            MenuFlag::Normal,
-            move |_| Message::Quit,
-        )
-}
-
-fn popup(_: &mut Frame, event: Event) -> bool {
-    match event {
-        Event::Push => match app::event_mouse_button() {
-            app::MouseButton::Right => {
-                crate::menu().popup();
-                true
-            }
-            _ => false,
-        },
-        _ => false,
-    }
-}
-
-fn frame(value: &str) -> Frame {
-    let mut element = Frame::default().with_label(value);
-    element.set_label_size(60);
-    element
-}
-
-fn cairobutton() -> Button {
-    let mut element = Button::default();
-    element.super_draw(false);
-    element.draw(|button| {
-        draw::draw_rect_fill(
-            button.x(),
-            button.y(),
-            button.w(),
-            button.h(),
-            Color::from_u32(0xfdf6e3),
-        );
-        let mut surface = ImageSurface::create(Format::ARgb32, button.w(), button.h())
-            .expect("Couldn’t create surface");
-        crate::draw_surface(&mut surface, button.w(), button.h());
-        if !button.value() {
-            cairo_blur::blur_image_surface(&mut surface, 20);
-        }
-        surface
-            .with_data(|surface| {
-                RgbImage::new(surface, button.w(), button.h(), ColorDepth::Rgba8)
-                    .unwrap()
-                    .draw(button.x(), button.y(), button.w(), button.h());
-            })
-            .unwrap();
-        draw::set_draw_color(Color::Black);
-        draw::set_font(Font::Helvetica, app::font_size());
-        if !button.value() {
-            draw::draw_rbox(
-                button.x() + 1,
-                button.y() + 1,
-                button.w() - 6,
-                button.h() - 6,
-                15,
-                true,
-                Color::White,
-            );
-            draw::draw_text2(
-                &button.label(),
-                button.x() + 1,
-                button.y() + 1,
-                button.w() - 6,
-                button.h() - 6,
-                Align::Center,
-            );
-        } else {
-            draw::draw_rbox(
-                button.x() + 1,
-                button.y() + 1,
-                button.w() - 4,
-                button.h() - 4,
-                15,
-                true,
-                Color::White,
-            );
-            draw::draw_text2(
-                &button.label(),
-                button.x() + 1,
-                button.y() + 1,
-                button.w() - 4,
-                button.h() - 4,
-                Align::Center,
-            );
-        }
-    });
-    element
-}
-
-fn draw_surface(surface: &mut ImageSurface, w: i32, h: i32) {
-    let ctx = Context::new(surface).unwrap();
+fn draw_box_with_alpha(rect: &mut Frame) {
+    let ctx = unsafe { Context::from_raw_none(fltk::app::cairo::cc() as _) };
+    let (r, g, b) = rect.color().to_rgb();
     ctx.save().unwrap();
-    let corner_radius = h as f64 / 10.0;
-    let radius = corner_radius / 1.0;
-    let degrees = std::f64::consts::PI / 180.0;
-
-    ctx.new_sub_path();
-    ctx.arc(w as f64 - radius, radius, radius, -90. * degrees, 0.0);
-    ctx.arc(
-        w as f64 - radius,
-        h as f64 - radius,
-        radius,
-        0.0,
-        90. * degrees,
-    );
-    ctx.arc(
-        radius,
-        h as f64 - radius,
-        radius,
-        90. * degrees,
-        180. * degrees,
-    );
-    ctx.arc(radius, radius, radius, 180. * degrees, 270. * degrees);
+    ctx.move_to(rect.x() as f64, rect.y() as f64);
+    ctx.line_to((rect.x() + rect.w()) as f64, rect.y() as f64);
+    ctx.line_to((rect.x() + rect.w()) as f64, (rect.y() + rect.h()) as f64);
+    ctx.line_to(rect.x() as f64, (rect.y() + rect.h()) as f64);
     ctx.close_path();
-
-    ctx.set_source_rgba(150.0 / 255.0, 150.0 / 255.0, 150.0 / 255.0, 40.0 / 255.0);
-    ctx.set_line_width(4.);
+    ctx.set_source_rgba(
+        r as f64 / 255.0,
+        g as f64 / 255.0,
+        b as f64 / 255.0,
+        100.0 / 255.0,
+    );
     ctx.fill().unwrap();
     ctx.restore().unwrap();
+}
+
+pub fn cairowidget(x: i32, y: i32, w: i32, h: i32, label: &str) -> Frame {
+    let mut element = Frame::new(x, y, w, h, None).with_label(label);
+    element.super_draw_first(false); // required for windows
+    element.draw(draw_box_with_alpha);
+    element
 }
