+++ demos/cairo/src/model/mod.rs
@@ -1,18 +1,12 @@
 pub struct Model {
-    value: u8,
+    pub state: [bool; 3],
 }
 
 impl Model {
     pub fn default() -> Self {
-        Self { value: 0u8 }
+        Self { state: [true; 3] }
     }
-    pub fn inc(&mut self) {
-        self.value = self.value.saturating_add(1);
-    }
-    pub fn dec(&mut self) {
-        self.value = self.value.saturating_sub(1);
-    }
-    pub fn value(&self) -> String {
-        self.value.to_string()
+    pub fn change(&mut self, idx: usize) {
+        self.state[idx] = !self.state[idx];
     }
 }
