+++ demos/cairo_shadow_button/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+authors = ["Mohammed Alyousef <may642_2000@hotmail.com>"]
+name = "cairo_button"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+flemish = { path = "../../" }
+cairo-rs = "0.18"
+cairo-blur = "^0.1"
