+++ demos/cairo_shadow_button/README.md
@@ -1,0 +1,5 @@
+# cairo shadow button
+
+This shows how to create a rounded button with a blur effect using cairo.
+
+![img](https://github.com/fltk-rs/demos/blob/master/cairo_shadow_button/assets/scrot.png)
