+++ demos/csv/README.md
@@ -1,0 +1,6 @@
+# CSV
+
+Custom drawing of CSV data. Uses CSV and Serde.
+
+![img](https://github.com/fltk-rs/demos/blob/master/csv/assets/csv.gif)
+
